Merged  2D Representation (#1)
This pull request adds a 2D representation of the map and player. The player is able to move around the map and pan to change its POV.

It also features a representation of where the players FOV collides with a wall (the yellow square is the center of the FOV). The FOV is changeable via the header file.